requirements.txt
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cssselect==1.2.0
distro==1.9.0
fastapi==0.112.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.15.4
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
jieba3k==0.35.1
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
lxml==5.3.0
lxml_html_clean==0.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
multidict==6.0.5
narwhals==1.4.2
newspaper3k==0.2.8
nltk==3.8.1
numpy==2.0.1
openai==0.28.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==5.27.3
pyarrow==17.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
rpds-py==0.20.0
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.2
streamlit==1.37.1
tenacity==8.5.0
tinysegmenter==0.3
tldextract==5.1.2
toml==0.10.2
tornado==6.4.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
watchdog==4.0.2
yarl==1.9.4
